1.initialize.f90: 之前第96行无法make, 原因是浮点数变量精度不匹配，我通过修改各个变量的精度一致，能够成功make. variables.f90, power_nu.f90相关的也都有修改。 2.修改utilities下的Makefile, 只make ic.x, cicpower.x 其他的如gadget_snap_fof.x还无法正常make,原因是还没有加入unzipx（不压缩坐标的）的逻辑。如果在Makefile下注释掉#OPTIONS+=-DZIPX, 尝试make gadget_snap_fof.x会报错。

+++ neutrinos/basic_functions.f08
@@ -45,7 +45,7 @@ endfunction
 pure function z2str(z)
    character(:),allocatable :: z2str
    character(20) :: str
-   real,intent(in) :: z
+   real(8),intent(in) :: z
    write(str,'(f7.3)') z
    z2str=trim(adjustl(str))
 endfunction
